Refactored to use improved naming conventions
