--- src/Bonsai.ML.PCA/PCAModelType.cs
@@ -1,19 +1,0 @@
-﻿using Bonsai;
-using System;
-using System.ComponentModel;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reactive.Linq;
-using TorchSharp;
-using static TorchSharp.torch;
-using static TorchSharp.torch.linalg;
-
-namespace Bonsai.ML.PCA
-{
-    public enum PCAModelType
-    {
-        PCA,
-        ProbabilisticPCA,
-        OnlinePPCA
-    }
-}
