+++ src/Bonsai.ML.Pca.Torch/PcaModelType.cs
@@ -1,0 +1,8 @@
+﻿namespace Bonsai.ML.Pca.Torch;
+
+public enum PcaModelType
+{
+    Pca,
+    ProbabilisticPca,
+    OnlineProbabilisticPca
+}
